--- a/docs/source/conf.py
+++ b/docs/source/conf.py
@@ -81,6 +81,7 @@ class CustomGalleryExampleSortKey:
                 "approximate_mode.py",
                 "sampling.py",
                 "parallel_decoding.py",
+                "custom_frame_mappings.py",
             ]
         else:
             assert "examples/encoding" in self.src_dir