… test/test_ops.py