This PR adds the code that will become the `custom_frame_mappings` tutorial. 
Currently, 2 benchmarks are run using a short video (~3 minutes) and a long video (~13 minutes) :
* Initializing VideoDecoders 
* Decoding 10 frames from 10 videos

Additionally, some fixes / optimizations were added to `custom_frame_mappings` logic in `SingleStreamDecoder.cpp` to improve the performance.

The page takes 1 minute 22 seconds to execute.

Benchmark results: 
<details> 

#### Compare performance of initializing VideoDecoder with custom_frame_mappings vs seek_modes

Running benchmarks on short_video.mp4
Creating a VideoDecoder object with custom_frame_mappings:
med = 7.68ms +- 20.83
Creating a VideoDecoder object with seek_mode='exact':
med = 17.26ms +- 14.54

Running benchmarks on long_video.mp4
Creating a VideoDecoder object with custom_frame_mappings:
med = 23.70ms +- 3.39
Creating a VideoDecoder object with seek_mode='exact':
med = 60.47ms +- 7.30

#### Decode frames with custom_frame_mappings vs exact seek_mode

Running benchmarks on short_video.mp4
Decoding frames with custom_frame_mappings:
med = 34.21ms +- 11.42
Decoding frames with seek_mode='exact':
med = 39.69ms +- 3.37

Running benchmarks on long_video.mp4
Decoding frames with custom_frame_mappings:
med = 50.53ms +- 8.38
Decoding frames with seek_mode='exact':
med = 83.95ms +- 17.38

</details>